Rename 'analytics' package to 'tracker'
