+++ code/testapp/src/main/java/com/adobe/edge/media/testapp/tracker/MediaAnalyticsProvider.java
@@ -9,7 +9,7 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.edge.media.testapp.analytics;
+package com.adobe.edge.media.testapp.tracker;
 
 import android.util.Log;
 import com.adobe.edge.media.testapp.Configuration;
