+++ code/testappkotlin/src/main/java/com/adobe/edge/media/testappkotlin/tracker/MediaAnalyticsProvider.kt
@@ -9,7 +9,7 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.edge.media.testappkotlin.analytics
+package com.adobe.edge.media.testappkotlin.tracker
 
 import android.util.Log
 import com.adobe.edge.media.testappkotlin.Configuration
